proctorforge/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
redis==5.1.1
websockets==12.0
httpx==0.27.2
reportlab==4.2.2
docker==7.1.0
anthropic==0.34.0
uuid6==2024.7.10
supabase>=2.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
